Feature ASAA-6 [SnippetDetails]
- Improving SnippetDetails.kt with more functionality

+++ domain/src/main/java/com/bottlerocketstudios/brarchitecture/domain/models/SnippetComment.kt
@@ -1,13 +1,13 @@
 package com.bottlerocketstudios.brarchitecture.domain.models
 
 class SnippetComment(
-    val id: Long?,
+    val id: Int?,
     val created: String?,
     val updated: String?,
     val content: SnippetCommentContent?,
     val user: User?,
     val deleted: Boolean?,
-    val parentId: Long?,
+    val parentId: Int?,
     val childrenComments: MutableList<SnippetComment?>,
     val links: Links? = null,
     val type: String?,
